:sparkles: Added dynamic calculation of swap distances.
As the coupling map of the atoms changes with each shuttling move operations, the distances are no longer pre-computed but calculated on the fly using breadth first search and then cached. A move operation resets all distances to uncomputed.
The cost function was also updated to support the now new possibliity of infinite distances.